Wide char support for commandLineArgs
I was unsure what the desired solution was here. Python3 will only talk in
wchar_t*, but this string needs to be handed over to QApplication,
which expects char*. I have made the wchar_t* vector only appear in Python3
and it simply converts the string across to the char* variety so the rest
of the code can continue to work.